requirements.txt
transformers>=4.57.0
qwen-vl-utils>=0.0.14
torch>=2.7.0
fastapi
uvicorn
pymilvus
pillow
matplotlib
modelscope
accelerate
requests
